refactor editor & fix serialization
